requirements.txt
albumentations==1.3.1
comet_ml==3.33.8
coremltools==6.3.0
Flask==2.3.2
GitPython==3.1.32
GitPython==3.1.32
ipython==8.14.0
matplotlib==3.7.2
mss==9.0.1
nncf==2.5.0
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.15.1
onnxsim==0.4.33
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
openvino==2023.0.1
paddle==1.0.2
pafy==0.5.5
pandas==2.0.3
Pillow==10.0.0
Pillow==7.0.0
Pillow==10.0.0
psutil==5.9.5
pycocotools==2.0.6
python-dotenv==1.0.0
PyYAML==6.0.1
Requests==2.31.0
roboflow==1.1.2
scipy==1.11.1
seaborn==0.12.2
setuptools==68.0.0
setuptools==45.2.0
tensorflow==2.13.0
tensorflow_datasets==4.9.2
tensorflowjs==4.10.0
tensorrt==8.6.1
tflite_runtime==2.13.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
tritonclient==2.36.0
ultralytics==8.0.145
x2paddle==1.4.1
